Phase 5: Legacy Cleanup - Remove hass.py and finalize modularization
- Removed app/hass.py (all functionality now in app/api/ modules)
- Updated all imports from app.hass to app.api modules
- Moved summarize_domain to app/api/entities.py
- Extracted prompts to app/prompts.py
- Reduced server.py from 1027 to 821 lines
- All tests passing (386 tests)
- Coverage: 80% overall
- All linting checks passing

Closes #54
