+++ tests/conftest.py
@@ -121,12 +121,11 @@ def mock_httpx_client():
         yield mock_client
 
 
-# Patch app.hass.get_client and app.core.client.get_client
+# Patch app.core.client.get_client and all API modules
 @pytest.fixture(autouse=True)
 def mock_get_client(mock_httpx_client):
     """Mock the get_client function to return our mock client."""
     with (
-        patch("app.hass.get_client", return_value=mock_httpx_client),
         patch("app.core.client.get_client", return_value=mock_httpx_client),
         patch("app.api.automations.get_client", return_value=mock_httpx_client),
         patch("app.api.entities.get_client", return_value=mock_httpx_client),
